backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
yfinance==0.2.51
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
google-generativeai==0.8.4
httpx==0.28.1
pydantic==2.10.4
cachetools==5.5.1
